Fixed indexing type.

+++ Processors/Z80/Z80.hpp
@@ -426,7 +426,7 @@ template <class T> class Processor {
 			size_t destination = 0;
 			for(size_t c = 0; c < 256; c++) {
 				target.instructions[c] = &target.all_operations[destination];
-				for(int t = 0; t < lengths[c];) {
+				for(size_t t = 0; t < lengths[c];) {
 					// Skip zero-length bus cycles.
 					if(table[c][t].type == MicroOp::BusOperation && table[c][t].machine_cycle.length == 0) {
 						t++;
